- Experimental Branch Merge #504
git-svn-id: svn://localhost/mosa@1405 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Mosa/HelloWorld/CMOS.cs
@@ -27,7 +27,7 @@ namespace Mosa.HelloWorld
 			Native.Nop();
 			Native.Nop();
 			Native.Nop();
-			byte result = Native.In(0x71);
+			byte result = (byte)Native.In(0x71);
 			//Native.Sti();
 			return result;
 		}
